requirements.txt
pydantic~=2.4.2
aiomysql~=0.2.0
fastapi~=0.103.2
PyYAML~=6.0.1
uvicorn~=0.23.2
psutil~=5.9.5
schedule~=1.2.1